# api/swagger.yaml
definitions:
  songs.createSongDTO:
    properties:
      group:
        type: string
      song:
        type: string
    type: object
  songs.songDTO:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        items:
          type: string
        type: array
    type: object
  songs.updateSongDTO:
    properties:
      group:
        type: string
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  title: Effective Mobile Song Library Service
  version: 0.0.1
paths:
  /songs:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Last song id
        in: query
        name: lastId
        type: integer
      - description: Filter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/songs.songDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get songs
      tags:
      - songs
    post:
      consumes:
      - application/json
      parameters:
      - description: Song data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/songs.createSongDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/songs.songDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create song
      tags:
      - songs
  /songs/{songId}:
    delete:
      parameters:
      - description: Song id
        in: path
        name: songId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete song
      tags:
      - songs
    patch:
      consumes:
      - application/json
      parameters:
      - description: Song id
        in: path
        name: songId
        required: true
        type: integer
      - description: Song data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/songs.updateSongDTO'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update song
      tags:
      - songs
  /songs/{songId}/lyrics:
    get:
      parameters:
      - description: Song id
        in: path
        name: songId
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get lyrics
      tags:
      - songs
swagger: "2.0"